requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx==0.26.0
python-multipart==0.0.6
greenlet==3.0.3
beautifulsoup4==4.12.3
